python_learn/frame_api/requirements.txt
boto3==1.12.22
botocore==1.15.22
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
cymysql==0.9.14
DateTime==4.3
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Executor==0.9.3
Flask-HTTPAuth==3.3.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.17.4
Pillow==7.0.0
pip==20.0.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
python-dateutil==2.8.1
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.2
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scandir==1.10.0
scipy==1.4.1
setuptools==41.6.0
six==1.13.0
smart-open==1.10.0
SQLAlchemy==1.2.18
torch==1.3.1
urllib3==1.25.8
Werkzeug==0.16.0
wheel==0.33.6
WTForms==2.2.1
zope.interface==4.7.1
